Add state for checking opening section visibility
I've added a state, hook, and observer for checking that at least 25% of
the Splash component is on screen. The reason I did this is because I
want the title h1 on the Header to expand and show my full name when the
opening section is in view. I also want it to condense into the shorter
version whenever the user scrolls away from that section.