Add state for checking initial load for animating
The functionality of the previous commit was to check if the opening
section was visible. With this commit, if the section is visible, the
header title will expand, but if the user scrolls away, it condenses.
I wanted there to be a delay on the expanding animation so that it
triggers after the Splash animation was done. However, I didn't like the
delay that was still present when users would re-trigger the animation
upon the Splash component becoming visible again, so I set the
initialLoad state for running the animation once with a delay.