Add functionality to Navigator using refs
Each main section has a ref attached to it now, and the anchor tags in
the navigator menu scrollIntoView these sections when clicked. I added
a scroll-margin-top to account for the fixed height of the header. The
reason I went this route instead of simply adding 'href="#sectionId"' to
each anchor, was because I didn't like how the address line gets changed
to reveal the ids of these elements. Also, the scrollToSection function
will scroll to the top of the page if a prop isn't provided. This is
intended to be useful later, as I plan on making the header title into
an anchor tag for bringing the user back up to the top.