Fix nav menu width and make header scroll to top
When the header is clicked it now scrolls the user to the top of the
page. In the previous commit, I intended on using the scrollToSection
function for this, but realized it would be simpler to do it separately
as it was a simple, one line function. I also fixed the width of the
nav menu, which was just slightly unaligned with the nav menu button.